Add files via upload
This GitHub repository contains an Email Classifier designed to differentiate between spam and non-spam (ham) emails using the Naive Bayes algorithm. The classifier is implemented without utilizing the scikit-learn library, relying solely on the numpy and pandas libraries for data manipulation and computation.
By leveraging this repository, users can gain insights into the workings of the Naive Bayes algorithm and its application in spam email classification tasks. Contributions, feedback, and collaborations are encouraged to enhance the functionality, efficiency, and usability of the classifier.